+++ src/store/sagas/auth.ts
@@ -32,7 +32,6 @@ export function* authUserSaga(action: AuthUser) {
   try {
     const response = yield axiosInstance
       .post(url, authData)
-    // dispatch
     const userId = response.data.localId;
     const idToken = response.data.idToken;
     const expiresInSeconds = response.data.expiresIn;
@@ -49,5 +48,27 @@ export function* authUserSaga(action: AuthUser) {
     yield put(checkAuthTimeout(expiresInSeconds));
   } catch (error) {
     yield put(authFail(error.response.data.error));
+  }
+}
+
+export function* authInitSaga () {
+  const userId = localStorage.getItem('userId');
+  const idToken = localStorage.getItem('idToken');
+  const expirationTime = parseInt(localStorage.getItem('expirationTime')!, 10);
+
+  if (idToken) {
+    const expirationTimeFromNow = expirationTime - new Date().getTime();
+    const expirationTimeFromNowInSeconds = expirationTimeFromNow / 1000;
+    if (expirationTimeFromNow <= 0) {
+      yield put(authInitiateLogout());
+    } else {
+      yield put(
+        authSuccess({
+          userId,
+          idToken,
+        }),
+      );
+      yield put(checkAuthTimeout(expirationTimeFromNowInSeconds));
+    }
   }
 }
